requirements.txt
llama-index
psycopg2-binary
python-dotenv
streamlit
streamlit-chat
sqlalchemy
black
isort
streamlit-supabase-auth
supabase
uvicorn
fastapi
cryptography